Refactor auth screens
